tratando de crear directorios y modularizar
