+++ juego/src/main.rs
@@ -1,215 +1,33 @@
 use std::io;
-use std::cmp::Ordering;
-use std::ptr::null;
-use rand::Rng;
-use std::collections::HashMap;
+//use std::cmp::Ordering;
+//use std::ptr::null;
+//use rand::Rng;
+//use std::collections::HashMap;
 use crossterm::{execute, terminal::{ClearType, Clear}}; // biblioteca para limpiar terminal
 
+mod habitacion;
+mod nivel;
+
+use crate::nivel::{Nivel, crear_niveles};
+use crate::habitacion::{/*Habitacion,*/ Posicion, imprimir_habitacion, inicializar_habitaciones_nivel};
+
 const NIVELES_POR_JUEGO: usize = 1;// POR AHORA 1 PERO SON 5
-const HABITACIONES_POR_NIVEL: usize = 2; // POR AHORA 2 PERO SON 5
-const PUERTAS_POR_HABITACION: usize = 4;
-const INDICE_JUGADOR: usize = 0;
 
 struct Juego{
     niveles: [Nivel; NIVELES_POR_JUEGO],
 }
 
-struct Nivel{
-    habitaciones: [Habitacion; HABITACIONES_POR_NIVEL],
-}
-
-struct Habitacion{
-    dimension_x : u32,
-    dimension_y : u32,
-    puertas: [Puerta; PUERTAS_POR_HABITACION],
-    jugadores: Vec<Jugador>,
-    enemigos: Vec<Enemigo>,
-    objetos_suelo : Vec<TipoObjeto>,
-}
-
-struct Puerta{
-    posicion: Posicion,
-    desde_hab: u32,
-    hasta_hab: u32,
-}
-
-enum Entidades{
-    Jugador(Jugador),
-    Objetos(TipoObjeto),
-    Enemigos(Enemigo),
-}
-
-struct Posicion{
-    x: u32,    
-    y: u32,
-}
-
-enum TipoObjeto {
-    Arma(Arma),
-    Pocion(Pocion),
-    Armadura(Armadura),
-}
-
-struct Arma{
-    daño: u32,
-    prob_crit: u32,
-    punteria: u32,
-}
-struct Pocion{
-    duracion: u32,
-    funcionalidad: fn(),
-}
-
-enum TipoPocion{
-    Vida,
-    Fuerza,
-    PielDeHierro,
-    Invisibilidad,
-    Punteria,
-    Esquiva,
-}
-
-impl Pocion{
-    fn vida(jugador: &mut Jugador){
-        jugador.atributos.salud_actual += 40;
-    }
-
-    fn fuerza(jugador: &mut Jugador){
-        jugador.atributos.daño += 10;
-    }
-
-    fn pieldehierro(jugador: &mut Jugador){
-        jugador.atributos.armadura += 50;
-    }
-
-    fn invisibilidad(jugador: &mut Jugador){
-        jugador.atributos.invisible = true;
-    }
-
-    fn punteria(jugador: &mut Jugador){
-        jugador.atributos.punteria += 20;
-    }
-
-    fn esquiva(jugador: &mut Jugador){
-        jugador.atributos.esquiva += 30;
-    }
-}
-
-struct Armadura{
-    armadura: u32,
-    esquiva: u32,
-}
-
-struct Atributos{
-    nombre: String,
-    posicion: Posicion,
-    salud_max: u32,
-    salud_actual: u32,
-    daño: u32,
-    prob_crit: u32,
-    armadura: u32,
-    punteria: u32,
-    esquiva: u32,
-    invisible: bool,
-}
-
-struct Jugador{
-    atributos: Atributos,
-    arma_equipada: Arma,
-    armadura_equipada: Armadura,
-    inventario: Vec<TipoObjeto>, 
-}
-
-struct Enemigo{
-
-}
-
-impl Jugador {
-    fn recoger_objeto(&mut self, objeto: TipoObjeto) {
-        self.inventario.push(objeto)
-    }
-    fn usar_pocion(&mut self, indice_pocion: usize){
-        
-    }
-}
-
 fn crear_juego() -> Juego{
     
-    let mut habitaciones = [
-        Habitacion {
-            dimension_x: 10,
-            dimension_y: 10,
-            puertas: [
-                Puerta {
-                    posicion: Posicion { x: 1, y: 2 }, 
-                    desde_hab: 0, 
-                    hasta_hab: 1, 
-                },
-                Puerta {
-                    posicion: Posicion { x: 5, y: 7 }, 
-                    desde_hab: 0, 
-                    hasta_hab: 2, 
-                },
-                Puerta {
-                    posicion: Posicion { x: 8, y: 3 }, 
-                    desde_hab: 1, 
-                    hasta_hab: 0, 
-                },
-                Puerta {
-                    posicion: Posicion { x: 2, y: 8 }, 
-                    desde_hab: 2, 
-                    hasta_hab: 0, 
-                },
-            ],
-            jugadores: Vec::new(),
-            enemigos: Vec::new(),
-            objetos_suelo: Vec::new(),
-        },
-        Habitacion {
-            dimension_x: 10,
-            dimension_y: 10,
-            puertas: [
-                Puerta {
-                    posicion: Posicion { x: 1, y: 2 }, 
-                    desde_hab: 0, 
-                    hasta_hab: 1, 
-                },
-                Puerta {
-                    posicion: Posicion { x: 5, y: 7 }, 
-                    desde_hab: 0, 
-                    hasta_hab: 2,
-                },
-                Puerta {
-                    posicion: Posicion { x: 8, y: 3 }, 
-                    desde_hab: 1, 
-                    hasta_hab: 0, 
-                },
-                Puerta {
-                    posicion: Posicion { x: 2, y: 8 },
-                    desde_hab: 2, 
-                    hasta_hab: 0, 
-                },
-            ],
-            jugadores: Vec::new(),
-            enemigos: Vec::new(),
-            objetos_suelo: Vec::new(),
-        },
-    ];
-
-    let mut nivel = Nivel {
-        habitaciones,
-    };
-       
+    let mut niveles: crear_niveles();
 
     let mut juego = Juego {
-        niveles: [nivel],
+        niveles: [Nivel],
     };
 
     juego
 
-
 }
-
 
 
 /* 
@@ -219,7 +37,7 @@ struct Trampa{
 }
 */
 
-fn es_critico(prob_critico: u32) -> bool{
+/*fn es_critico(prob_critico: u32) -> bool{
     let chance_minima_necesaria: u32 = rand::thread_rng().gen_range(1..=100);
     if prob_critico >= chance_minima_necesaria {
         return true;
@@ -240,12 +58,13 @@ fn golpe(chance_de_golpe: u32, prob_critico: u32) -> u32{
         return 2;
     }
 
-}
+}*/
 
-/* Combate entre dos entidades, disminuye el daño de ser necesario.
+/* 
+    Combate entre dos entidades, disminuye el daño de ser necesario.
     Tiene en cuenta la punteria y el esquive.
- */
-fn combate(mut ent_1: Atributos, mut ent_2: Atributos){
+*/
+/*fn combate(mut ent_1: Atributos, mut ent_2: Atributos){
     let chance_de_golpe: u32 = ent_1.punteria - ent_2.esquiva;
 
     let resul_golpe  = golpe(chance_de_golpe, ent_1.prob_crit);
@@ -256,7 +75,7 @@ fn combate(mut ent_1: Atributos, mut ent_2: Atributos){
         2 => println!("{} fallo su ataque!", ent_1.nombre),
         _ => return,
     }
-}
+}*/
 
 fn movimiento(pos_actual: &mut Posicion, direccion: char){
     match direccion{
@@ -266,19 +85,6 @@ fn movimiento(pos_actual: &mut Posicion, direccion: char){
         'd' => pos_actual.x += 1,
         _ => return,
     };
-}
-
-// Esto genera una posicion aleatoria dentro de las dimensiones de la habitacion que recibe para colocar los objetos y enemigos al inicializar
-fn generar_pos_en_hab(habitacion: &mut Habitacion) -> Posicion{
-    // let x_gen: u32 = rand::thread_rng().gen_range(1..=habitacion.dimension_x);
-    // let y_gen: u32 = rand::thread_rng().gen_range(1..=habitacion.dimension_y);
-
-    let posicion: Posicion = Posicion {
-        x : rand::thread_rng().gen_range(1..=habitacion.dimension_x),
-        y : rand::thread_rng().gen_range(1..=habitacion.dimension_y),
-    };
-
-    return posicion;
 }
 
 /* 
@@ -297,68 +103,12 @@ fn recoger_pocion(pocion: Pocion, jugador: Jugador){
 }
 */
 
-fn esta_tocando_puerta(habitacion: Habitacion) -> bool{
-    return true;
-}
 
 // Hace que el jugador pase de una habitacion a otra
-fn pasar_de_habitacion(jugador: Jugador, habitacion1: &mut Habitacion, habitacion2: &mut Habitacion){
+/*fn pasar_de_habitacion(jugador: Jugador, habitacion1: &mut Habitacion, habitacion2: &mut Habitacion){
     habitacion1.jugadores.retain(|j| j.atributos.nombre != jugador.atributos.nombre);
     habitacion2.jugadores.push(jugador);
-}
-
-fn generar_dimensiones_hab(habitacion: &mut Habitacion) -> &mut Habitacion{
-    habitacion.dimension_x = rand::thread_rng().gen_range(4..=10);
-    habitacion.dimension_y = rand::thread_rng().gen_range(4..=10);
-
-    return habitacion;
-}
-
-fn generar_puertas(habitacion: &mut Habitacion) -> &mut Habitacion {
-    for mut puertas in &mut habitacion.puertas{
-           
-        let puerta = Puerta { 
-            posicion: Posicion { x: 1, y: 2 },  // falta esa logica para que sea random
-            desde_hab: 0, 
-            hasta_hab: 1,
-        };
-    }
-    habitacion
-}
-
-// Inicializa las habitaciones de un nivel, con sus objetos, enemigos, tamaño, etc
-fn inicializar_habitaciones_nivel(nivel: &mut Nivel){
-    for mut habitacion in &mut nivel.habitaciones{
-        habitacion = generar_dimensiones_hab(habitacion);
-        habitacion = generar_puertas(habitacion); // hace falta que returnee habitaciones?
-        habitacion = inicializar_jugador(habitacion) 
-    }
-}
-
-fn imprimir_habitacion(habitacion: &Habitacion) {
-
-
-    println!("Dimensiones: {}x{}", habitacion.dimension_x, habitacion.dimension_y);
-    let mut matriz: Vec<Vec<String>> = vec![vec!["-".to_string(); habitacion.dimension_x as usize]; habitacion.dimension_y as usize]; // creamos una matriz de tamaño x y 
- 
-
-    for jugador in &habitacion.jugadores {
-        if jugador.atributos.posicion.x < habitacion.dimension_x && jugador.atributos.posicion.y < habitacion.dimension_y {
-            matriz[jugador.atributos.posicion.y as usize][jugador.atributos.posicion.x as usize] = "J".to_string();
-        }
-    }
-
-    for fila in matriz.iter() {
-        for celda in fila.iter() {
-        print!("{}  ", celda);
-        }
-        println!();
-    }
-
-
-   // println!("Puertas: {:?}", habitacion.puertas);
-   // println!("Jugadores: {:?}", habitacion.jugadores);
-}
+}*/
 
 fn imprimir_tablero(juego: &mut Juego){
 
@@ -369,81 +119,35 @@ fn imprimir_tablero(juego: &mut Juego){
             println!();
         }
     }
-
 }
-
 
 
 fn imprimir_mapa(juego: &mut Juego){
     execute!(std::io::stdout(), Clear(ClearType::All)).unwrap(); // borrar pantalla
     
-    
     /*  
     println!(nivel , armadura, posciones)
     println! objetos inventario
     */ 
-
     imprimir_tablero(juego);
 
 }
 
 
-fn inicializar_jugador(habitacion: &mut Habitacion) -> &mut Habitacion{
-
-    let jugador = Jugador {
-        atributos: Atributos {
-            nombre: "String".to_string(), // esto es correcto??
-            posicion: generar_pos_en_hab( habitacion),
-            salud_max: 100,
-            salud_actual: 100,
-            daño: 100,
-            prob_crit: 0,
-            armadura: 100,
-            punteria: 0,
-            esquiva: 0,
-            invisible: false,
-        },
-        arma_equipada: Arma {
-            daño: 0,
-            prob_crit: 0,
-            punteria: 0,
-        },
-        armadura_equipada: Armadura {
-            armadura: 0,
-            esquiva: 0,
-        },
-        inventario: Vec::new(), // Inicializa el inventario como un vector vacío
-    };
- 
-    habitacion.jugadores.push(jugador);
-
-    return habitacion;
-
-   
-}
-
-
-
 // Inicializa el juego, con todos sus niveles, habitaciones, etc y jugados / enemigos y objetos
 fn inicializar_juego(juego: &mut Juego){
+
+    let cantidad_habitaciones = 4;
    
     for nivel in juego.niveles.iter_mut() {
-            inicializar_habitaciones_nivel(nivel);
+            inicializar_habitaciones_nivel(cantidad_habitaciones);
     }
     
     //enemigos --> generar_pos_en_hab
     //inicializar mapas
 
     // inicializar objetos
-
 }
-
-/*fn es_posicion_valida(posicion: Posicion, habitacion: &Habitacion) -> bool {
-    let max_x = habitacion.dimension_x - 1;
-    let max_y = habitacion.dimension_y - 1;
-    posicion.x >= 0 && posicion.x <= max_x && posicion.y >= 0 && posicion.y <= max_y
-}*/
-
 
 fn recibir_movimiento(juego: &mut Juego)-> bool{
